requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Markdown>=3.5.0
PyYAML>=6.0.0
